22/07/16 last commit (리뷰 작성 구현)

+++ app/src/main/java/com/zibro/fooddeliveryapp/widget/adapter/viewholder/review/ReviewViewHolder.kt
@@ -36,7 +36,7 @@ class ReviewViewHolder(
 
             reviewTitleText.text = model.title
             reviewText.text = model.description
-            ratingBar.rating = model.grade.toFloat()
+            ratingBar.rating = model.grade
         }
     }
     override fun bindViews(model: RestaurantReviewModel, adapterListener: AdapterListener) = Unit
